--- src/main/kotlin/no/nav/familie/ba/infotrygd/feed/rest/dto/FeedMeldingDto.kt
@@ -1,7 +1,0 @@
-package no.nav.familie.ba.infotrygd.feed.rest.dto
-
-data class FeedMeldingDto(
-        val elementer: List<FeedElement>,
-        val inneholderFlereElementer: Boolean,
-        val tittel: String
-)
