--- src/main/kotlin/no/nav/familie/ba/infotrygd/feed/rest/dto/Type.kt
@@ -1,6 +1,0 @@
-package no.nav.familie.ba.infotrygd.feed.rest.dto
-
-enum class Type {
-    BA_Vedtak_v1,
-    BA_Foedsel_v1;
-}
