+++ src/main/kotlin/no/nav/familie/ef/infotrygd/feed/rest/dto/ElementMetadata.kt
@@ -1,0 +1,7 @@
+package no.nav.familie.ef.infotrygd.feed.rest.dto
+
+import java.time.LocalDateTime
+
+data class ElementMetadata(
+    val opprettetDato: LocalDateTime // TODO: NB K9 bruker java.time.OffSetDateTime her
+)
