+++ src/main/kotlin/no/nav/familie/ef/infotrygd/feed/rest/dto/FeedElement.kt
@@ -1,0 +1,8 @@
+package no.nav.familie.ef.infotrygd.feed.rest.dto
+
+data class FeedElement(
+        val innhold: Innhold,
+        val metadata: ElementMetadata,
+        val sekvensId: Int,
+        val type: Type
+)
