+++ src/main/kotlin/no/nav/familie/ef/infotrygd/feed/rest/dto/FeedMeldingDto.kt
@@ -1,0 +1,7 @@
+package no.nav.familie.ef.infotrygd.feed.rest.dto
+
+data class FeedMeldingDto(
+        val elementer: List<FeedElement>,
+        val inneholderFlereElementer: Boolean,
+        val tittel: String
+)
