+++ src/main/kotlin/no/nav/familie/ef/infotrygd/feed/rest/dto/Innhold.kt
@@ -1,0 +1,9 @@
+package no.nav.familie.ef.infotrygd.feed.rest.dto
+
+import java.time.LocalDate
+
+interface Innhold
+
+data class InnholdVedtak(val datoStartNyBA : LocalDate, val fnrStoenadsmottaker: String) : Innhold
+
+data class InnholdFødsel(val fnrBarn: String) : Innhold
