+++ src/main/kotlin/no/nav/familie/ef/infotrygd/feed/rest/dto/Type.kt
@@ -1,0 +1,6 @@
+package no.nav.familie.ef.infotrygd.feed.rest.dto
+
+enum class Type {
+    BA_Vedtak_v1,
+    BA_Foedsel_v1;
+}
