--- src/test/kotlin/no/nav/familie/ba/infotrygd/DevLauncher.kt
@@ -1,13 +1,0 @@
-package no.nav.familie.ba.infotrygd
-
-import org.springframework.boot.SpringApplication
-import org.springframework.boot.autoconfigure.SpringBootApplication
-
-@SpringBootApplication(scanBasePackages = ["no.nav.familie.ba.infotrygd.feed"])
-class DevLauncher
-
-fun main(args: Array<String>) {
-    val springApp = SpringApplication(DevLauncher::class.java)
-    springApp.setAdditionalProfiles("dev")
-    springApp.run(*args)
-}
