--- src/test/kotlin/no/nav/familie/ba/infotrygd/feed/database/DbContainerInitializer.kt
@@ -1,35 +1,0 @@
-package no.nav.familie.ba.infotrygd.feed.database
-
-import org.springframework.boot.test.util.TestPropertyValues
-import org.springframework.context.ApplicationContextInitializer
-import org.springframework.context.ConfigurableApplicationContext
-import org.springframework.core.env.Profiles
-import org.testcontainers.containers.PostgreSQLContainer
-
-class DbContainerInitializer : ApplicationContextInitializer<ConfigurableApplicationContext> {
-
-    override fun initialize(applicationContext: ConfigurableApplicationContext) {
-        // Only start Postgres when not running in CI
-        if (!applicationContext.environment.acceptsProfiles(Profiles.of("ci"))) {
-            postgres.start()
-            TestPropertyValues.of(
-                    "spring.datasource.url=${postgres.jdbcUrl}",
-                    "spring.datasource.username=${postgres.username}",
-                    "spring.datasource.password=${postgres.password}"
-            ).applyTo(applicationContext.environment)
-        }
-    }
-
-    companion object {
-        // Lazy because we only want it to be initialized when accessed
-        private val postgres: KPostgreSQLContainer by lazy {
-            KPostgreSQLContainer("postgres:11.1")
-                    .withDatabaseName("databasename")
-                    .withUsername("postgres")
-                    .withPassword("test")
-        }
-    }
-}
-
-// Hack needed because testcontainers use of generics confuses Kotlin
-class KPostgreSQLContainer(imageName: String) : PostgreSQLContainer<KPostgreSQLContainer>(imageName)
